toasty/samplers.py: move the sampler functions here
It makes sense to have these in their own module, I think.
