+++ docs/api/toasty.samplers.healpix_fits_file_sampler.rst
@@ -1,0 +1,6 @@
+healpix_fits_file_sampler
+=========================
+
+.. currentmodule:: toasty.samplers
+
+.. autofunction:: healpix_fits_file_sampler
