--- docs/api/toasty.tile.normalizer.rst
@@ -1,6 +1,0 @@
-normalizer
-==========
-
-.. currentmodule:: toasty.tile
-
-.. autofunction:: normalizer
