+++ docs/api/toasty.toast.depth2tiles.rst
@@ -1,0 +1,6 @@
+depth2tiles
+===========
+
+.. currentmodule:: toasty.toast
+
+.. autofunction:: depth2tiles
