+++ docs/api/toasty.toast.gen_wtml.rst
@@ -1,0 +1,6 @@
+gen_wtml
+========
+
+.. currentmodule:: toasty.toast
+
+.. autofunction:: gen_wtml
