+++ docs/api/toasty.toast.iter_corners.rst
@@ -1,0 +1,6 @@
+iter_corners
+============
+
+.. currentmodule:: toasty.toast
+
+.. autofunction:: iter_corners
