+++ docs/api/toasty.toast.iter_tiles.rst
@@ -1,0 +1,6 @@
+iter_tiles
+==========
+
+.. currentmodule:: toasty.toast
+
+.. autofunction:: iter_tiles
