+++ docs/api/toasty.toast.minmax_tile_in_range.rst
@@ -1,0 +1,6 @@
+minmax_tile_in_range
+====================
+
+.. currentmodule:: toasty.toast
+
+.. autofunction:: minmax_tile_in_range
