+++ toasty/__init__.py
@@ -4,12 +4,9 @@
 from __future__ import absolute_import, division, print_function
 
 from .toast import (
-    cartesian_sampler,
     depth2tiles,
     gen_wtml,
-    healpix_sampler,
     iter_corners,
     iter_tiles,
     minmax_tile_in_range,
-    normalizer,
 )
